Configures development environment
Sets up the development environment for the project,
including Ruby LSP configuration, Dockerfile updates,
and database setup scripts.

Key changes:
- Adds Ruby LSP support for enhanced code editing.
- Configures port forwarding for the Ruby LSP.
- Initializes Postgres and configures database creation and seeding.
- Updates Dockerfile to use a non-root user.
- Implements API v2 search functionality with CORS support and improved search logic.
- Updates Rubocop config.

+++ .devcontainer/postSetup.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/zsh
+#!/usr/bin/env zsh
 LOG_FILE="/workspaces/greenwood-kiosk/.devcontainer/postSetup.log"
 if [[ -e ".env" ]]; then
     echo -e "\e[33m.env file found: skipping creation\e[0m" | tee $LOG_FILE
@@ -13,16 +13,40 @@ secret_key_base=$(rails secret) | tee -a $LOG_FILE
 export DEVISE_SECRET_KEY=$devise_secret | tee -a $LOG_FILE
 export SECRET_KEY_BASE=$secret_key_base | tee -a $LOG_FILE
 
-echo -e "\e[33mSetting up the database...\e[0m" | tee -a $LOG_FILE
-if [[ -e "config/database.yml" ]]; then
-    echo -e "\e[33mdatabase.yml file found: skipping creation\e[0m" | tee -a $LOG_FILE
+echo -e "\e[33mChecking if database exists...\e[0m" | tee -a $LOG_FILE
+
+DB_EXISTS=$(rails db:exists 2>&1 || echo "false")
+
+if echo "$DB_EXISTS" | grep -q "Database does not exist"; then
+    SCHEMA_EXISTS=$(rails db:schema_exists 2>&1 || echo "false")
+    DB_EMPTY=$(rails db:is_empty 2>&1 || echo "false")
+
+    echo -e "\e[33mDatabase does not exist, creating and setting up...\e[0m" | tee -a $LOG_FILE
+
+    if [[ -e "config/database.yml" ]]; then
+        echo -e "\e[33mdatabase.yml file found: skipping creation\e[0m" | tee -a $LOG_FILE
+    else
+        echo -e "\e[33mCreating database.yml file...\e[0m" | tee -a $LOG_FILE
+        cp config/database.example.yml config/database.yml | tee -a $LOG_FILE
+    fi
+
+    echo -e "\e[33mCreating the database...\e[0m" | tee -a $LOG_FILE
+    rails db:create | tee -a $LOG_FILE
+    if echo "$DB_EXISTS" | grep -q "Schema does not exist"; then
+        echo -e "\e[33mLoading schema...\e[0m" | tee -a $LOG_FILE
+        rails db:schema:load | tee -a $LOG_FILE
+    else 
+        echo -e "\e[33mSchema already exists: skipping loading\e[0m" | tee -a $LOG_FILE
+    fi
+    if echo "$DB_EMPTY" | grep -q  "Database is empty"; then
+        echo -e "\e[33mSeeding the DB...\e[0m" | tee -a $LOG_FILE
+        rails db:seed | tee -a $LOG_FILE
+    else
+        echo -e "\e[33mDatabase is not empty: skipping seeding\e[0m" | tee -a $LOG_FILE
+    fi
 else
-    echo -e "\e[33mCreating database.yml file...\e[0m" | tee -a $LOG_FILE
-    cp config/database.example.yml config/database.yml | tee -a $LOG_FILE
+    echo -e "\e[33mDatabase already exists: skipping creation and seeding\e[0m" | tee -a $LOG_FILE
 fi
-echo -e "\e[33mCreating the database...\e[0m" | tee -a $LOG_FILE
-rails db:create | tee -a $LOG_FILE
-echo -e "\e[33mLoading schema...\e[0m" | tee -a $LOG_FILE
-rails db:schema:load | tee -a $LOG_FILE
-echo -e "\e[33mSeeding the DB...\e[0m" | tee -a $LOG_FILE
-rails db:seed | tee -a $LOG_FILE
+
+bundle config set --local path 'vendor/bundle' && bundle install
+chmod +x /workspaces/greenwood-kiosk/.devcontainer/lsp-wrapper.sh
